python_core/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "linucast"
version = "0.1.0"
description = "AI-powered virtual camera enhancement for Linux"
authors = ["Linucast Team <team@linucast.org>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "linucast" }]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.1"
torchvision = "^0.15.1"
onnxruntime = "^1.16.0"
opencv-python = "^4.8.0"
mediapipe = "^0.10.0"
PyQt6 = "^6.5.0"
numpy = "^1.24"
pyyaml = "^6.0"
pillow = "^10.0.0"
insightface = "^0.7.3"
onnx = "^1.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"

[tool.poetry.scripts]
linucast = "linucast.__main__:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
